chore: rename files to keep the layer structure
