+++ docker-compose.yml
@@ -1,6 +1,6 @@
 version: '2'
 services:
-  couch_db:
+  users_db:
     image: klaemo/couchdb:2.0.0
     ports:
       - '3000:5984'
@@ -8,9 +8,10 @@ services:
   auth_ms:
     build: .
     environment:
-      PORT: '5984'
-      ROOT_URL: 'couch_db'
+      DB_PORT: '5984'
+      DB_ROOT_URL: 'users_db'
+      SECRET: 'ec6607e21b57ac2ccc5836754893ed33' # Put your own secret!
     ports:
       - '4005:4005'
     links:
-      - couch_db
+      - users_db
