+++ package.json
@@ -3,11 +3,11 @@
   "private": true,
   "version": "0.0.1",
   "description": "User authentication microservice",
-  "main": "./src/index.js",
+  "main": "./src/resources.js",
   "scripts": {
     "eslint": "eslint src test",
     "eslint-fix": "npm run eslint -- --fix",
-    "start": "node src/index.js",
+    "start": "node src/resources.js",
     "test": "npm run test-mocha && npm run eslint",
     "test-cov": "istanbul cover node_modules/.bin/_mocha -- --require should --reporter dot --recursive && codecov",
     "test-travis": "istanbul cover node_modules/.bin/_mocha --report lcovonly -- --require should --reporter dot --recursive && codecov",
