+++ src/models/auth.js
@@ -5,9 +5,9 @@ const status = require('http-status');
 const jwt = require('jsonwebtoken');
 
 // CouchDB url
-const PORT = process.env.PORT || 5984;
-const ROOT_URL = process.env.ROOT_URL || 'localhost';
-const url = `http://${ROOT_URL}:${PORT}/`;
+const DB_PORT = process.env.DB_PORT || 5984;
+const DB_ROOT_URL = process.env.DB_ROOT_URL || 'localhost';
+const url = `http://${DB_ROOT_URL}:${DB_PORT}/`;
 const dbUrl = url + 'blinkbox_users';
 
 const secret = process.env.SECRET || 'blinkbox is cool';
